rename misleading config folder
